+++ link-move/src/main/java/com/nhl/link/move/writer/TargetPropertyWriterService.java
@@ -1,15 +1,11 @@
 package com.nhl.link.move.writer;
 
 import com.nhl.link.move.LmRuntimeException;
+import com.nhl.link.move.runtime.cayenne.CayenneCrossVersionBinaryCompat;
 import com.nhl.link.move.runtime.cayenne.ITargetCayenneService;
 import org.apache.cayenne.di.Inject;
-import org.apache.cayenne.map.DbAttribute;
 import org.apache.cayenne.map.ObjEntity;
-import org.apache.cayenne.reflect.AttributeProperty;
-import org.apache.cayenne.reflect.ClassDescriptor;
-import org.apache.cayenne.reflect.PropertyVisitor;
-import org.apache.cayenne.reflect.ToManyProperty;
-import org.apache.cayenne.reflect.ToOneProperty;
+import org.apache.cayenne.reflect.*;
 
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -35,7 +31,7 @@ public class TargetPropertyWriterService implements ITargetPropertyWriterService
     }
 
     private <T> TargetPropertyWriterFactory<T> createWriterFactory(Class<T> type) {
-        
+
         ObjEntity entity = targetCayenneService.entityResolver().getObjEntity(type);
         if (entity == null) {
             throw new LmRuntimeException("Java class " + type.getName() + " is not mapped in Cayenne");
@@ -70,9 +66,7 @@ public class TargetPropertyWriterService implements ITargetPropertyWriterService
             }
         });
 
-        for (DbAttribute pk : entity.getDbEntity().getPrimaryKeys()) {
-            writerFactory.initPkWriter(pk);
-        }
+        CayenneCrossVersionBinaryCompat.pkAttributes(entity.getDbEntity()).forEach(writerFactory::initPkWriter);
 
         return writerFactory;
     }